refactor: rename folder store Snackbar
