--- src/shared/common/store/SnackBar/types/Snackbar.types.ts
@@ -1,12 +1,0 @@
-import { AlertColor, SnackbarProps } from "@mui/material";
-
-export type TypeSnackBar = "SNACKBAR_OPEN" | "SNACKBAR_CLOSE"
-
-export interface ISnackBarState extends SnackbarProps {
-    severety: AlertColor;
-}
-
-export interface ISnackbarAction {
-    type: TypeSnackBar;
-    payload: ISnackBarState;
-}
