+++ src/shared/components/Layout/Snackbar/Snackbar.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { useSelector } from "react-redux";
 import useSnackbar from "shared/common/hook/useSnackbar";
-import { ISnackBarState } from "shared/common/store/SnackBar/types/Snackbar.types";
+import { ISnackbarState } from "shared/common/store/Snackbar/types/Snackbar.types";
 import SnackbarView from "./SnackbarView";
 
 const THREE_SECONDS_AUTO_HIDE_DURATION = 3000;
@@ -9,7 +9,7 @@ const THREE_SECONDS_AUTO_HIDE_DURATION = 3000;
 const Snackbar: React.FC = () => {
 	const { snackbarClose } = useSnackbar();
 	const snackbarProps = useSelector(
-		(state: { snackBar: ISnackBarState }) => state.snackBar
+		(state: { snackbar: ISnackbarState }) => state.snackbar
 	);
 
 	const defineAutoHideDuration =
